test(api/v2): extract huma helpers into shared test file
